hv: Handle holes in GSI i.e. Global System Interrupt for multiple IO-APICs
MADT is used to specify the GSI base for each IO-APIC and the number of
interrupt pins per IO-APIC is programmed into Max. Redir. Entry register of
that IO-APIC.

On platforms with multiple IO-APICs, there can be holes in the GSI space.
For example, on a platform with 2 IO-APICs, the following configuration has
a hole (from 24 to 31) in the GSI space.

IO-APIC 1: GSI base - 0, number of pins - 24
IO-APIC 2: GSI base - 32, number of pins - 8

This patch also adjusts the size for variables used to represent the total
number of IO-APICs on the system from uint16_t to uint8_t as the ACPI MADT
uses only 8-bits to indicate the unique IO-APIC IDs.

Tracked-On: #4151
Acked-by: Eddie Dong <eddie.dong@Intel.com>

+++ hypervisor/boot/acpi_base.c
@@ -209,14 +209,14 @@ local_parse_madt(struct acpi_table_madt *madt, uint32_t lapic_id_array[MAX_PCPU_
 	return pcpu_num;
 }
 
-static uint16_t
+static uint8_t
 ioapic_parse_madt(void *madt, struct ioapic_info *ioapic_id_array)
 {
 	struct acpi_madt_ioapic *ioapic;
 	struct acpi_table_madt *madt_ptr;
 	void *first, *end, *iterator;
 	struct acpi_subtable_header *entry;
-	uint16_t ioapic_idx = 0U;
+	uint8_t ioapic_idx = 0U;
 
 	madt_ptr = (struct acpi_table_madt *)madt;
 
@@ -261,9 +261,9 @@ uint16_t parse_madt(uint32_t lapic_id_array[MAX_PCPU_NUM])
 	return ret;
 }
 
-uint16_t parse_madt_ioapic(struct ioapic_info *ioapic_id_array)
+uint8_t parse_madt_ioapic(struct ioapic_info *ioapic_id_array)
 {
-	uint16_t ret = 0U;
+	uint8_t ret = 0U;
 	struct acpi_table_rsdp *rsdp = NULL;
 
 	rsdp = get_rsdp();
